created files for modules. Moved game methods from stat_tracker to game module. Deleted helper ethods modoule and appended to end of games module.
