+++ lib/league_module.rb
@@ -1,0 +1,4 @@
+module League
+
+
+end
