+++ lib/season_module.rb
@@ -1,0 +1,3 @@
+module Season
+
+end
