+++ lib/team_module.rb
@@ -1,0 +1,5 @@
+module Team
+
+
+
+end
